respondd-module-airtime: Move common netlink query code to separate function
